refactor(tests): reduce duplication in test files below 5% threshold
- Extract shared `getResultText` + `createMockIntegrationClient` helpers
  into `__tests__/helpers/integration-test-utils.ts`; use them in
  `tso-tools.integration.test.ts` and `uss-tools.integration.test.ts`
  (both describe blocks)
- Extract JWT test helpers (`requestUrl`, `b64url`, `signJwt`,
  `postMcpLocal`) into `__tests__/helpers/jwt-test-utils.ts`; shared by
  `bearer-jwt.test.ts`, `http-transport-jwt.test.ts`, and
  `keycloak-http-jwt.e2e.test.ts`
- Replace "wait for server socket" boilerplate with the existing
  `waitUntil()` helper in `extension-client.test.ts` (4 call sites)
- Convert `native-backend.test.ts` listMembers / readDataset /
  searchInDataset / listDatasets tests to use the existing
  `createBackendWithClient()` factory (10 call sites)

Reduces test duplication from 49 clones (6.36%) to 35 clones (4.62%).
